requirements.txt
pydrobert-speech
torch>=1.10